+++ lib/announcement/models/announcement_dto/configuration.dart
@@ -5,33 +5,27 @@ part 'configuration.g.dart';
 
 @JsonSerializable()
 class Configuration {
+
+  const Configuration({this.pointsPerLike, this.pointsPerVideo});
+
+  factory Configuration.fromJson(Map<String, dynamic> json) => _$ConfigurationFromJson(json);
   @JsonKey(name: 'points_per_like')
   final int? pointsPerLike;
   @JsonKey(name: 'points_per_video')
   final int? pointsPerVideo;
 
-  const Configuration({this.pointsPerLike, this.pointsPerVideo});
-
   @override
-  String toString() {
-    return 'Configuration(pointsPerLike: $pointsPerLike, pointsPerVideo: $pointsPerVideo)';
-  }
-
-  factory Configuration.fromJson(Map<String, dynamic> json) {
-    return _$ConfigurationFromJson(json);
-  }
+  String toString() => 'Configuration(pointsPerLike: $pointsPerLike, pointsPerVideo: $pointsPerVideo)';
 
   Map<String, dynamic> toJson() => _$ConfigurationToJson(this);
 
   Configuration copyWith({
     int? pointsPerLike,
     int? pointsPerVideo,
-  }) {
-    return Configuration(
+  }) => Configuration(
       pointsPerLike: pointsPerLike ?? this.pointsPerLike,
       pointsPerVideo: pointsPerVideo ?? this.pointsPerVideo,
     );
-  }
 
   @override
   bool operator ==(Object other) {