+++ lib/notification/utils/notifications_utils.dart
@@ -53,6 +53,6 @@ abstract class NotificationsUtils {
   ) =>
       notifications
           .where(
-              (n) => n.type != null && isNotificationAllowedForWebById(n.type!))
+              (n) => n.type != null && isNotificationAllowedForWebById(n.type!),)
           .toList();
 }