+++ lib/utils/text_button.dart
@@ -39,6 +39,7 @@ class TextButtonStoyco extends StatelessWidget {
     this.fontWeight = FontWeight.w400,
     this.boxShadow,
     this.loadingIndicatorSize = 24.0,
+    this.textStyle,
   });
 
   final String text;
@@ -51,6 +52,7 @@ class TextButtonStoyco extends StatelessWidget {
   final FontWeight fontWeight;
   final List<BoxShadow>? boxShadow;
   final double loadingIndicatorSize;
+  final TextStyle? textStyle;
 
   @override
   Widget build(BuildContext context) {
@@ -65,15 +67,16 @@ class TextButtonStoyco extends StatelessWidget {
             )
           : Text(
               text,
-              style: TextStyle(
-                fontFamily: 'Akkurat_Pro',
-                fontSize: fontSize,
-                fontWeight: fontWeight,
-                height: 1.19,
-                letterSpacing: 0,
-                color: const Color(0xFFFAFAFA),
-                decoration: TextDecoration.none,
-              ),
+              style: textStyle ??
+                  TextStyle(
+                    fontFamily: 'Akkurat_Pro',
+                    fontSize: fontSize,
+                    fontWeight: fontWeight,
+                    height: 1.19,
+                    letterSpacing: 0,
+                    color: const Color(0xFFFAFAFA),
+                    decoration: TextDecoration.none,
+                  ),
             ),
     );
     return SizedBox(
@@ -132,6 +135,7 @@ class OutlinedTextButtonStoyco extends StatelessWidget {
     this.backgroundColor,
     this.fontSize = 16,
     this.fontWeight = FontWeight.w400,
+    this.textStyle,
   });
 
   final String text;
@@ -142,6 +146,7 @@ class OutlinedTextButtonStoyco extends StatelessWidget {
   final Color? backgroundColor;
   final double fontSize;
   final FontWeight fontWeight;
+  final TextStyle? textStyle;
 
   @override
   Widget build(BuildContext context) {
@@ -152,15 +157,16 @@ class OutlinedTextButtonStoyco extends StatelessWidget {
             )
           : Text(
               text,
-              style: TextStyle(
-                fontFamily: 'Akkurat_Pro',
-                fontSize: fontSize,
-                fontWeight: fontWeight,
-                height: 1.19,
-                letterSpacing: 0,
-                color: const Color(0xFFFAFAFA),
-                decoration: TextDecoration.none,
-              ),
+              style: textStyle ??
+                  TextStyle(
+                    fontFamily: 'Akkurat_Pro',
+                    fontSize: fontSize,
+                    fontWeight: fontWeight,
+                    height: 1.19,
+                    letterSpacing: 0,
+                    color: const Color(0xFFFAFAFA),
+                    decoration: TextDecoration.none,
+                  ),
             ),
     );
     return SizedBox(